requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai==0.8.3
chromadb
sentence-transformers
nomic
rank-bm25
numpy
scikit-learn
ragas
evaluate
fastapi
uvicorn
pydantic
python-dotenv
pandas
requests
tqdm
python-multipart
structlog
python-json-logger
pytest
black
flake8
mypy
einops
streamlit
pypdf
nltk
langchain-ollama
langchain-chroma
nest_asyncio
